requirements.txt
pillow
asgiref==3.5.0
coverage==6.3.2
crispy-tailwind==0.5.0
Django==3.1.4
django-crispy-forms==1.14.0
django-environ
black
celery
redis
django-celery-results
djangorestframework
markdown
django-filter
djangorestframework-simplejwt